Moved async opensearch providers
